The textbox work never ends...
(Hopefully) largely finalized the implementation of textboxes. Changed the style to remove the symbols and instead just have them be in the text. made the space prompt change depending on the circumstance, with three possible variations (skip, continue, exit). Implemented the cutscene, with the ability to show/hide character sprites and names.