File Restructuring, Game Restructuring, Scene Manager, etc.
This commit is the beginning of the general game restructuring that will help make the game more modular.
To start I made a game scene that will act as the central part of our game and never get unloaded. Levels and UI will instead be loaded in the screen_holder which is what the player will see. This will allow us to separate levels from UI so it's not constantly getting destroyed.

The SceneManager plays into this structure as it is what loads and unloads the new scene and the old scene into the screen_holder. To use this new manager I created a DoorButton which is similar to a normal button but upon pressing it will call the SceneManager and load in the new scene. I used this to replace the start button on the main menu.

I changed the File System slightly to better categorize what files go where. Entities replaced Scenes. Entities are all the physical things we use to fill up levels and the game such as the player, NPC's, boxes, buildings, props, etc. I added the Game Folder which holds everything pertinent to the structure of the game such as UI, Levels, Loading Screen, etc.